Refactor exceptions
