+++ src/main/java/seedu/address/cashier/logic/parser/EditCommandParser.java
@@ -9,7 +9,7 @@ import static seedu.address.util.CliSyntax.PREFIX_QUANTITY;
 import seedu.address.cashier.logic.commands.EditCommand;
 import seedu.address.cashier.logic.commands.exception.InsufficientAmountException;
 import seedu.address.cashier.logic.commands.exception.NotANumberException;
-import seedu.address.cashier.logic.exception.NegativeQuantityException;
+import seedu.address.cashier.logic.commands.exception.NegativeQuantityException;
 import seedu.address.cashier.logic.parser.exception.ParseException;
 import seedu.address.cashier.model.Model;
 import seedu.address.cashier.model.exception.NoSuchItemException;
