+++ src/test/java/seedu/address/cashier/logic/AddCommandParserTest.java
@@ -17,4 +17,9 @@ public class AddCommandParserTest {
     public void parse_allFieldsPresent_success() {
 
     }
+
+    @Test
+    public void parse_compulsoryFieldMissing_failure() {
+
+    }
 }
